+++ src/controllers/students.controller.js
@@ -2,6 +2,10 @@ import {pool} from "../db.js"
 
 export const createStudent = async(req,res)=>{
 try {
+    const [row,info] = await pool.query("SELECT * FROM users WHERE email = ? AND password = ?",[req.body.emailUser,req.body.password])
+
+    if(row.lenght > 0){
+
     const data = await pool.query("INSERT INTO students (nombre,apellidos,correo,especialidad,grado,grupo) VALUES(?,?,?,?,?,?)",
     [req.body.nombre,req.body.apellidos,req.body.correo,req.body.especialidad,req.body.grado,req.body.grupo])
     const group = await pool.query("SELECT * FROM classrooms WHERE especialidad = ? AND grado = ? AND grupo = ?",
@@ -21,7 +25,7 @@ rows.map((e)=>{
     )
     
     res.send(data[0])
-
+    }
 } catch (error) {
     res.send(error)
 }
@@ -29,9 +33,13 @@ rows.map((e)=>{
 
 export const getStudents = async(req,res)=>{
     try{
+        const [row,info] = await pool.query("SELECT * FROM users WHERE email = ? AND password = ?",[req.body.emailUser,req.body.password])
+
+        if(row.lenght > 0){
         const data = await pool.query("SELECT * FROM students WHERE especialidad = ? AND grado = ? AND grupo = ?",[req.params.especialidad,req.params.grado,req.params.grupo])
         if(data[0].length === 0){
             res.send([])
+        }
         }else{
             res.send(data[0])
         }
@@ -46,8 +54,12 @@ res.send(error)
 
 export const getStudent = async(req,res)=>{
     try{
+        const [row,info] = await pool.query("SELECT * FROM users WHERE email = ? AND password = ?",[req.body.emailUser,req.body.password])
+
+        if(row.lenght > 0){
         const data = await pool.query("SELECT * FROM students WHERE id = ?",[req.params.id])
         res.send(data[0])
+        }
     }catch(error){
 res.send(error)
     }
@@ -58,10 +70,14 @@ res.send(error)
 
 export const deleteStudent = async(req,res)=>{
     try{
+        const [row,info] = await pool.query("SELECT * FROM users WHERE email = ? AND password = ?",[req.body.emailUser,req.body.password])
+
+        if(row.lenght > 0){
         await pool.query("SET FOREIGN_KEY_CHECKS=0")
         const data = await pool.query("DELETE FROM students WHERE id = ? ",[req.params.id])
         await pool.query("SET FOREIGN_KEY_CHECKS=1")
         res.send(data[0])
+        }
     }catch(error){
 res.send(error)
     }
@@ -72,9 +88,12 @@ res.send(error)
 
 export const updateStudent = async(req,res)=>{
     try{
-    
+        const [row,info] = await pool.query("SELECT * FROM users WHERE email = ? AND password = ?",[req.body.emailUser,req.body.password])
+
+    if(row.lenght > 0){
         const data = await pool.query("UPDATE students SET nombre = ?, apellidos = ?, correo = ? WHERE id = ?",[req.body.nombre,req.body.apellidos,req.body.correo,req.params.id])
         res.send(data[0])
+    }
     }catch(error){
 res.send(error)
     }
