+++ src/controllers/tasks.controller.js
@@ -2,15 +2,19 @@ import {pool} from "../db.js"
 
 export const createTask = async (req,res)=>{
     try {
-      await pool.query("INSERT INTO tasks (name,rate,grade,groupTask,area) VALUES (?,?,?,?,?)",[req.body.nombre,req.body.rate,req.body.grade,req.body.group,req.body.area])
+const [row,info] = await pool.query("SELECT * FROM users WHERE email = ? AND password = ?",[req.body.emailUser,req.body.password])
 
-        await req.body.alumnosTask.forEach(async(element) => {
-    await pool.query("INSERT INTO tasks_students (name,rate,final_rate,task_for) VALUES (?,?,?,?)",[req.body.nombre,req.body.rate,req.body.finalRate,element.id])
+if(row.lenght > 0){
+  await pool.query("INSERT INTO tasks (name,rate,grade,groupTask,area) VALUES (?,?,?,?,?)",[req.body.nombre,req.body.rate,req.body.grade,req.body.group,req.body.area])
 
-        });
-        res.send("todo bien")
+  await req.body.alumnosTask.forEach(async(element) => {
+await pool.query("INSERT INTO tasks_students (name,rate,final_rate,task_for) VALUES (?,?,?,?)",[req.body.nombre,req.body.rate,req.body.finalRate,element.id])
 
-    } catch (error) {
+  });
+  res.send("todo bien")
+
+}
+} catch (error) {
         res.send(error)
     }
    
@@ -19,13 +23,18 @@ export const createTask = async (req,res)=>{
 
 export const deleteTask = async (req,res)=>{
   try {
-    await pool.query("DELETE FROM tasks WHERE id = ?",[req.body.id])
 
-      await req.body.alumnosTask.forEach(async(element) => {
-  await pool.query("DELETE FROM tasks_students WHERE name = ? AND task_for = ?",[req.body.nameTask,element.id])
+    const [row,info] = await pool.query("SELECT * FROM users WHERE email = ? AND password = ?",[req.body.emailUser,req.body.password])
 
-      });
-      res.send("todo bien")
+if(row.lenght > 0){
+  await pool.query("DELETE FROM tasks WHERE id = ?",[req.body.id])
+
+  await req.body.alumnosTask.forEach(async(element) => {
+await pool.query("DELETE FROM tasks_students WHERE name = ? AND task_for = ?",[req.body.nameTask,element.id])
+
+  });
+  res.send("todo bien")
+}
 
   } catch (error) {
       res.send(error)
@@ -37,8 +46,13 @@ export const deleteTask = async (req,res)=>{
 export const getTasks = async (req,res)=>{
    
     try {
+      const [row,info] = await pool.query("SELECT * FROM users WHERE email = ? AND password = ?",[req.body.emailUser,req.body.password])
+
+      if(row.lenght > 0){
         const data = await pool.query("SELECT * FROM tasks_students WHERE task_for = ?",[req.params.id])
         res.send(data[0])  
+      }
+        
     } catch (error) {
         res.send(error)
 
@@ -50,8 +64,13 @@ export const getTasks = async (req,res)=>{
 export const getTasksGroup = async (req,res)=>{
    
    try {
-    const data = await pool.query("SELECT * FROM tasks WHERE grade = ? AND groupTask = ? AND area = ?",[req.params.grade,req.params.group,req.params.area])
-    res.send(data[0])
+
+    const [row,info] = await pool.query("SELECT * FROM users WHERE email = ? AND password = ?",[req.body.emailUser,req.body.password])
+
+    if(row.lenght > 0){
+      const data = await pool.query("SELECT * FROM tasks WHERE grade = ? AND groupTask = ? AND area = ?",[req.params.grade,req.params.group,req.params.area])
+      res.send(data[0])
+    }
    } catch (error) {
     res.send(error)
    }
@@ -60,8 +79,14 @@ export const getTasksGroup = async (req,res)=>{
 
     export const changeRateTask = async (req,res)=>{
   try {
-    const data = await pool.query("UPDATE tasks_students SET final_rate = ? WHERE task_for = ? AND name = ?",[req.body.newRate,req.body.idStudent,req.body.taskName])
-    res.send(data)
+
+    
+    const [row,info] = await pool.query("SELECT * FROM users WHERE email = ? AND password = ?",[req.body.emailUser,req.body.password])
+
+    if(row.lenght > 0){
+      const data = await pool.query("UPDATE tasks_students SET final_rate = ? WHERE task_for = ? AND name = ?",[req.body.newRate,req.body.idStudent,req.body.taskName])
+      res.send(data)
+    }
   } catch (error) {
     res.send(error)
   }
@@ -71,16 +96,21 @@ export const getTasksGroup = async (req,res)=>{
     export const changeNameTask = async (req,res)=>{
       try {
 
-        const data = await pool.query("UPDATE tasks SET name = ? WHERE  id = ?",[req.body.newTaskName,req.body.idTask])
-       req.body.alumnosTask.forEach(async(element) => {
-        await pool.query("SET SQL_SAFE_UPDATES = 0")
+        
+    const [row,info] = await pool.query("SELECT * FROM users WHERE email = ? AND password = ?",[req.body.emailUser,req.body.password])
 
-        await pool.query("UPDATE tasks_students SET name = ? WHERE  task_for = ? AND name = ?",[req.body.newTaskName,element.id,req.body.nameTask])
-        })
-        await pool.query("SET SQL_SAFE_UPDATES = 1")
+    if(row.lenght > 0){
+      const data = await pool.query("UPDATE tasks SET name = ? WHERE  id = ?",[req.body.newTaskName,req.body.idTask])
+      req.body.alumnosTask.forEach(async(element) => {
+       await pool.query("SET SQL_SAFE_UPDATES = 0")
+
+       await pool.query("UPDATE tasks_students SET name = ? WHERE  task_for = ? AND name = ?",[req.body.newTaskName,element.id,req.body.nameTask])
+       })
+       await pool.query("SET SQL_SAFE_UPDATES = 1")
 
 
-        res.send(data)
+       res.send(data)
+    }
       } catch (error) {
         res.send(error)
       }
@@ -89,16 +119,20 @@ export const getTasksGroup = async (req,res)=>{
 
         export const changeRateTaskGroup = async (req,res)=>{
           try {
-    
-            const data = await pool.query("UPDATE tasks SET rate = ? WHERE  id = ?",[req.body.newRate,req.body.idTask])
-           req.body.alumnosTask.forEach(async(element) => {
-            await pool.query("SET SQL_SAFE_UPDATES = 0")
-    
-            await pool.query("UPDATE tasks_students SET rate = ? WHERE  task_for = ? AND name = ?",[req.body.newRate,element.id,req.body.nameTask])
-      
-            })
-    
-            res.send(data)
+
+            const [row,info] = await pool.query("SELECT * FROM users WHERE email = ? AND password = ?",[req.body.emailUser,req.body.password])
+
+            if(row.lenght > 0){
+              const data = await pool.query("UPDATE tasks SET rate = ? WHERE  id = ?",[req.body.newRate,req.body.idTask])
+              req.body.alumnosTask.forEach(async(element) => {
+               await pool.query("SET SQL_SAFE_UPDATES = 0")
+       
+               await pool.query("UPDATE tasks_students SET rate = ? WHERE  task_for = ? AND name = ?",[req.body.newRate,element.id,req.body.nameTask])
+         
+               })
+       
+               res.send(data)
+            }
           } catch (error) {
            res.send(error)
           }
@@ -112,8 +146,15 @@ export const getTasksGroup = async (req,res)=>{
 
     export const downloadExcel = async (req,res)=>{
       try {
-        const data = await pool.query("UPDATE tasks_students SET final_rate = ? WHERE task_for = ? AND name = ?",[req.body.newRate,req.body.idStudent,req.body.taskName])
-        res.send(data)
+
+        
+        const [row,info] = await pool.query("SELECT * FROM users WHERE email = ? AND password = ?",[req.body.emailUser,req.body.password])
+
+        if(row.lenght > 0){
+          const data = await pool.query("UPDATE tasks_students SET final_rate = ? WHERE task_for = ? AND name = ?",[req.body.newRate,req.body.idStudent,req.body.taskName])
+          res.send(data)
+        }
+       
       } catch (error) {
         res.send(error)
       }
